+++ archunit/src/main/java/com/tngtech/archunit/library/cycle_detection/CycleInternal.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.tngtech.archunit.library.dependencies;
+package com.tngtech.archunit.library.cycle_detection;
 
 import java.util.List;
 import java.util.Objects;
