+++ archunit/src/main/java/com/tngtech/archunit/library/cycle_detection/PrimitiveDataTypes.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.tngtech.archunit.library.dependencies;
+package com.tngtech.archunit.library.cycle_detection;
 
 import static com.google.common.base.Preconditions.checkState;
 import static java.util.Arrays.copyOf;
